fix(check2-findings): kernel-small 13条处置(12修+1文档)——P1: DataParameterBinders.FLOAT声明DOUBLE实取Float对齐getDouble/setDouble
- P1: FLOAT binder改getDouble/setDouble/(Double)（红:CCE Double cannot be cast to Float）,REAL保持自洽,AutoConvert包装路径随之自洽
- P2×6: ReflectClass.remove键统一getSignature+allDeclaredMethods分支纠正+删mergeList死代码/MarkdownSectionMerger summary改取sectionB（红:摘要丢失）/TableViewToMarkdownTableConverter补齐改padToColumns(报告建议setSize为截断语义不适用已注明)/PomModelMerger补else put+modules不继承（红×3）/CodeBlock.append复用buf（红:abc→ac）/getObjectConstructor非collection兜底LinkedHashMap（红:NPE）
- P3: SingleColumnRow.setObject抛ERR_DATASET_IS_READONLY(全仓核查仅读路径构造)/JdkJavaCompiler -source/-target取Runtime.version().feature()(不再固定1.8,红:var语法编译失败)+URI吞异常改ERR_JAVAC_INVALID_CLASS_NAME/KernelCliValidateCommand printStackTrace→LOG.error(异常对象入参保留堆栈)/requireField改传getFieldNames/RowNumberRecordInput javadoc类型约束(文档方案)
- 新增7测试类+2扩展;stash红11处; dataset 15+codegen 23+markdown 80+record-mapping 51+javac 7 tests绿 + 报告13/13标注 + plan346勾选
- 超范围发现(记plan346 follow-up未修): nop-kernel-cli pom -proc:only(7f8b955822笔误)致模块源码从不编译、测试从未执行

+++ nop-kernel/nop-dataset/src/main/java/io/nop/dataset/binder/DataParameterBinders.java
@@ -216,12 +216,12 @@ public class DataParameterBinders {
 
         @Override
         public Object getValue(IDataParameters params, int index) {
-            return params.getFloat(index);
+            return params.getDouble(index);
         }
 
         @Override
         public void setValue(IDataParameters params, int index, Object value) {
-            params.setFloat(index, (Float) value);
+            params.setDouble(index, (Double) value);
         }
     };
